api_gateway/requirements.txt
flask==2.3.2
flask-cors==3.0.10
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.7.0
pymysql==1.0.3
werkzeug==2.3.7
flask-sqlalchemy==3.0.5
flask-migrate==4.0.4
